move fact details to details package
