+++ app/src/main/java/com/example/testapplication/ui/details/FactDetailActivity.kt
@@ -1,4 +1,4 @@
-package com.example.testapplication.ui.listItems
+package com.example.testapplication.ui.details
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
