+++ app/src/main/java/com/example/testapplication/ui/listItems/MainActivity.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import androidx.core.view.isVisible
 import com.example.testapplication.R
 import com.example.testapplication.entity.model.CatData
+import com.example.testapplication.ui.details.FactDetailActivity
 import com.example.testapplication.utils.Params
 import io.reactivex.disposables.CompositeDisposable
 import kotlinx.android.synthetic.main.activity_main.*
